WIP: Minor fixes and to-be-replaced type parameter support
