+++ build.sbt
@@ -6,7 +6,7 @@ lazy val scalatest = Def.setting("org.scalatest" %%% "scalatest" % "3.2.16")
 lazy val specs2 = Def.setting("org.specs2" %%% "specs2-core" % "4.20.2")
 
 val commonSettings = Defaults.coreDefaultSettings ++ Seq(
-  scalaVersion := "3.3.1",
+  scalaVersion := "3.3.0",
   scalacOptions ++= Seq("-deprecation", "-unchecked", "-feature", "-release:8")
 )
