+++ shared/src/test/scala/org/scalamock/test/scalatest/BasicTest.scala
