+++ shared/src/test/scala/org/scalamock/test/scalatest/PathSpecTest.scala
